Rearrange a bit so that Angular stuff can be lazy-loaded
