+++ src/app/inventory/inventory.module.ts
@@ -1,14 +1,12 @@
 import { module } from 'angular';
 
-import { tagIconFilter, StoreItemComponent } from './dimStoreItem.directive';
 import { SimpleItemComponent } from './dimSimpleItem.directive';
-import { PercentWidth, percent } from './dimPercentWidth.directive';
+import { PercentWidth } from './dimPercentWidth.directive';
+import { tagIconFilter } from './dim-item-info';
 
 const mod = module('inventoryModule', [])
   .component('dimSimpleItem', SimpleItemComponent)
-  .component('dimStoreItem', StoreItemComponent)
   .directive('dimPercentWidth', PercentWidth)
-  .filter('tagIcon', tagIconFilter)
-  .filter('percent', () => percent);
+  .filter('tagIcon', tagIconFilter);
 
 export default mod.name;
