+++ src/app/move-popup/move-popup.module.ts
@@ -8,6 +8,7 @@ import { MoveItemPropertiesComponent } from './dimMoveItemProperties.directive';
 import { MovePopupComponent } from './dimMovePopup.directive';
 import { MoveLocationsComponent } from './move-locations.component';
 import { talentGridNodesFilter, TalentGridComponent } from './talent-grid.component';
+import { StarRatingComponent } from '../shell/star-rating/star-rating.component';
 import Sockets from './Sockets';
 import { ItemPopup } from './item-popup.directive';
 import { PressTip } from './press-tip.directive';
@@ -24,6 +25,7 @@ export default module('movePopupModule', [])
   .component('dimMoveLocations', MoveLocationsComponent)
   .component('dimMovePopup', MovePopupComponent)
   .component('dimTalentGrid', TalentGridComponent)
+  .component('starRating', StarRatingComponent)
   .component('sockets', react2angular(Sockets, ['item'], ['$scope']))
   .filter('talentGridNodes', () => talentGridNodesFilter)
   .filter('ammoTypeClass', () => ammoTypeClass)
